Moved mapping into own folder, namespace
